refactor(e2e): extract dismiss method tests into dismiss-dialog.spec.ts
Move parameterized Escape key and overlay click tests from
cancel-operation.spec.ts into a dedicated dismiss-dialog.spec.ts file.
