+++ host-frontend-root/frontend-src-root/tests/e2e/pages/rule-list/features/delete-rule/cancel-operation.spec.ts
@@ -1,4 +1,3 @@
-import type { Page } from '@playwright/test';
 import { expect, test } from 'tests/e2e/fixtures';
 import {
   assertNoConsoleErrors,
@@ -118,65 +117,4 @@ test.describe('ルール削除機能 - キャンセル操作', () => {
     // 7. Assert: コンソールエラーが発生していないことを確認
     assertNoConsoleErrors(consoleMessages);
   });
-
-  const dismissMethods = [
-    {
-      name: 'Escapeキー',
-      dismiss: async (rulesPage: Page) => {
-        await rulesPage.keyboard.press('Escape');
-      },
-    },
-    {
-      name: 'オーバーレイクリック',
-      dismiss: async (rulesPage: Page) => {
-        const overlay = rulesPage.locator(
-          '[data-testid="confirm-dialog-overlay"]'
-        );
-        await expect(overlay).toBeVisible();
-        await overlay.click({ position: { x: 10, y: 10 } });
-      },
-    },
-  ];
-
-  for (const { name, dismiss } of dismissMethods) {
-    test(`${name}でダイアログを閉じるとルールが残る`, async ({
-      page,
-      popupPage,
-      rulesPage,
-    }) => {
-      // コンソールエラー監視をセットアップ
-      const consoleMessages = setupConsoleErrorMonitoring(popupPage, rulesPage);
-
-      // 1. Arrange: ルールを保存
-      const testOldString = `${name}テスト`;
-      await saveRule(popupPage, page, {
-        oldString: testOldString,
-        newString: '置換後',
-      });
-
-      // 2. Arrange: ルール一覧ページをリロード
-      await reloadAndWaitForTable(rulesPage);
-
-      // 3. Act: ゴミ箱アイコンをクリック
-      await clickDeleteButton(rulesPage, 0);
-      await waitForConfirmDialog(rulesPage);
-
-      // 4. Act: ダイアログを閉じる
-      await dismiss(rulesPage);
-
-      // 5. Assert: ダイアログが閉じる
-      await waitForConfirmDialogClosed(rulesPage);
-
-      // 6. Assert: ルールが残っている
-      const count = await getRuleCount(rulesPage);
-      expect(count).toBe(1);
-
-      // 7. Assert: ルールの内容が変わっていない
-      const oldString = await getRuleOldString(rulesPage, 0);
-      expect(oldString).toBe(testOldString);
-
-      // 8. Assert: コンソールエラーが発生していないことを確認
-      assertNoConsoleErrors(consoleMessages);
-    });
-  }
 });
